Factor out contour line related codes
